move mapping options function into own utils file
